+++ R/useShinyalert.R
@@ -4,6 +4,9 @@
 #' \code{\link[shinyalert]{shinyalert}} function to work.\cr\cr
 #' You can call \code{useShinyalert()} from anywhere inside the UI.
 #'
+#' @param rmd Set this to \code{TRUE} if using \code{shinyalert}
+#' inside an interactive R markdown document. The YAML of the Rmd file must have
+#' \code{runtime: shiny}.
 #' @return Scripts that \code{shinyalert} requires that are automatically
 #' inserted to the app's \code{<head>} tag.
 #' @examples
@@ -26,12 +29,16 @@
 #' }
 #' @seealso \code{\link[shinyalert]{shinyalert}}
 #' @export
-useShinyalert <- function() {
+useShinyalert <- function(rmd = FALSE) {
+  stopifnot(rmd == TRUE || rmd == FALSE)
+
   shiny::addResourcePath("resources",
                          system.file("www", package = "shinyalert"))
 
+  insert_into_doc <- if (rmd) shiny::tagList else shiny::tags$head
+
   shiny::singleton(
-    shiny::tags$head(
+    insert_into_doc(
       shiny::tags$script(
         src = file.path("resources", "shared", "sweetalert-1.0.1",
                         "js", "sweetalert.min.js")